question_service/requirements.txt
asgiref==3.9.2
dj-database-url==2.1.0
Django==5.2.6
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
